separate server implementation into files
