+++ src/servers/swagger.ts
@@ -1,0 +1,15 @@
+  //   server.register(
+  //     swagger,
+  //     withRefResolver({
+  //       routePrefix: "/docs",
+  //       exposeRoute: true,
+  //       staticCSP: true,
+  //       openapi: {
+  //         info: {
+  //           title: "Fastify API",
+  //           description: "API for some products",
+  //           version,
+  //         },
+  //       },
+  //     })
+  //   );
